weather-bot/requirements.txt
httpx>=0.24.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
aiohttp>=3.8.0
schedule>=1.2.0
rich>=13.0.0
scipy>=1.11.0
